refactor(floating-window): pin ウィンドウ機構を session-log から汎用 feature に抽出
intent(floating-window): preview パネルにも session-log と同じドラッグ切り離しを載せるため、ドラッグ移動 / 8方位リサイズ / drag handoff / z 管理 / ビューポートクランプを共有シェル FloatingWindow + createFloatingWindows factory に切り出す
decision(floating-window): z カウンタは全 factory instance で共有 — plain fixed 要素同士の相対順はウィンドウ種類を跨ぐため、instance ごとに分けると種類を跨いだ bring-to-front が効かなくなる
decision(floating-window): ウィンドウ操作ボタンは actions slot + 共通 close をヘッダ右上の 1 グループに集約 — シェル / consumer のコード所有権境界が画面配置に漏れてボタンが行を跨いで散るのを契約で防ぐ
