+++ apps/renderer/src/features/floating-window/index.ts
@@ -1,0 +1,12 @@
+// floating-window feature の公開 API。
+//
+// 「一時 UI (popover / pane) の表示中コンテンツをドラッグで独立フローティングウィンドウへ
+// 昇格させる」機構が複数 feature にまたがる (session-log の pinned log / preview の
+// pinned preview) ため、汎用シェルと状態管理 factory を独立 feature として切り出している。
+// 何を pin するか (payload とヘッダ / 本文の描画) は各 consumer 側に閉じる。
+export { default as FloatingWindow } from "./FloatingWindow.vue";
+export {
+  createFloatingWindows,
+  type FloatingWindowState,
+  type PinDragHandoff,
+} from "./useFloatingWindows";
