+++ apps/renderer/src/features/floating-window/useFloatingWindows.ts
@@ -1,0 +1,96 @@
+/**
+ * pin されたフローティングウィンドウ群の状態管理 factory。
+ *
+ * pinned window は pin 元 (popover / pane) の表示状態と独立して存在し続けるため、
+ * consumer feature (session-log / preview) が module スコープで `createFloatingWindows<T>()`
+ * を 1 回だけ実行し、payload T を載せた singleton のウィンドウ列を得る。
+ *
+ * 位置は pin 元 (popover の box) の実測 rect、サイズは pin 元の本文 (スクロール面) の
+ * 実測を初期値として受け取る。popover がその場でフローティング化したように見せる視覚的
+ * 連続性のため。サイズを総高さでなく本文で受け渡すのは、pin 元とウィンドウでヘッダの
+ * 高さが違うため。総高さを引き継ぐと増えたヘッダ分だけ本文が食われて切れる。ウィンドウは
+ * mount 時に自分のヘッダ実測高を足して総高さを決める。以後、位置 (x / y) はドラッグと
+ * 左/上辺リサイズで更新されるが、サイズは store 上では初期値のまま不変で、リサイズの
+ * SSOT はリサイズハンドラが書く DOM の inline style に移る (FloatingWindow の doc 参照)。
+ *
+ * z カウンタは全 factory instance で共有する。種類の異なるウィンドウ (log / preview) も
+ * 同じ plain fixed のスタッキング文脈に並ぶため、カウンタを instance ごとに分けると
+ * 種類を跨いだ bring-to-front が効かなくなる。
+ */
+import { ref, type Ref } from "vue";
+
+export interface FloatingWindowState {
+  id: number;
+  x: number;
+  y: number;
+  /**
+   * 初期の本文 (スクロール面) サイズ (pin 元の本文実測。総サイズでない理由は
+   * モジュール docstring 参照)。mount 後はリサイズハンドラが inline style を上書きする。
+   */
+  bodyWidth: number;
+  bodyHeight: number;
+  z: number;
+}
+
+/**
+ * pin と同時にドラッグを開始する引き継ぎ情報。pin 元ヘッダのドラッグで pin する経路では、
+ * 掴んでいた要素が pin と同時に消える (unmount / hide) ため pointer capture を持ち越せない。
+ * pin() がこれを預かり、mount されたウィンドウが takeHandoff() で 1 回だけ消費して
+ * 同じ pointerId のドラッグとして継続する。
+ */
+export interface PinDragHandoff {
+  pointerId: number;
+  /** pointer からウィンドウ原点 (rect 左上) へのオフセット。 */
+  offsetX: number;
+  offsetY: number;
+}
+
+// bring-to-front の z 初期値。plain fixed 要素どうしの相対順にだけ効き、dialog / popover の
+// top layer は z-index に関係なく常に手前 (ArcadeLayer の doc 参照)。
+const Z_BASE = 30;
+
+// 全 factory instance で共有 (モジュール docstring 参照)。
+let zTop = Z_BASE;
+
+export function createFloatingWindows<T>() {
+  const windows = ref([]) as Ref<(T & FloatingWindowState)[]>;
+  let nextId = 0;
+  // pin() → mount → takeHandoff() が同期フラッシュ内で完結するため reactive にしない。
+  let pendingHandoff: ({ id: number } & PinDragHandoff) | undefined;
+
+  function pin(input: T & Omit<FloatingWindowState, "id" | "z">, handoff?: PinDragHandoff) {
+    const id = nextId++;
+    windows.value.push({ ...input, id, z: ++zTop });
+    if (handoff !== undefined) pendingHandoff = { id, ...handoff };
+  }
+
+  /** id 宛の drag handoff を 1 回だけ消費する。無ければ undefined。 */
+  function takeHandoff(id: number): PinDragHandoff | undefined {
+    if (pendingHandoff === undefined || pendingHandoff.id !== id) return undefined;
+    const { pointerId, offsetX, offsetY } = pendingHandoff;
+    pendingHandoff = undefined;
+    return { pointerId, offsetX, offsetY };
+  }
+
+  function close(id: number) {
+    windows.value = windows.value.filter((w) => w.id !== id);
+  }
+
+  function move(id: number, x: number, y: number) {
+    const win = windows.value.find((w) => w.id === id);
+    if (win === undefined) return;
+    win.x = x;
+    win.y = y;
+  }
+
+  function bringToFront(id: number) {
+    const win = windows.value.find((w) => w.id === id);
+    if (win === undefined) return;
+    // zTop は全種共有のため、他種のウィンドウが後から pin されていれば z !== zTop になり
+    // 正しく再前面化される
+    if (win.z === zTop) return;
+    win.z = ++zTop;
+  }
+
+  return { windows, pin, takeHandoff, close, move, bringToFront };
+}
