+++ apps/renderer/src/features/session-log/index.ts
@@ -8,6 +8,6 @@ export { expandAskMessages, parseSessionLog, type TranscriptEvent } from "@gozd/
 export { default as PinnedLogLayer } from "./PinnedLogLayer.vue";
 export { default as SessionLogDialog } from "./SessionLogDialog.vue";
 export { default as SessionLogMessageBody } from "./SessionLogMessageBody.vue";
-export { usePinnedLog, type PinDragHandoff } from "./usePinnedLog";
+export { usePinnedLog } from "./usePinnedLog";
 export { useSessionLogLive } from "./useSessionLogLive";
 export { useSessionLogViewer } from "./useSessionLogViewer";
